Protect against trying to display monster health bar in town
That was resulting in trying to dereference a null pointer and crashing
the game. When I merged the Infernity monster healt bar patch in
ade1413f6b7e59e1200ca03c7bbd6f50e02dceac I missed the fact that there's
only one method to draw view in DevilutionX now so this needs to use an
additional condition here.